Milestone 3
Added in proper enemy controller, a "guard" style enemy (he basically grabs a weapon and starts firing at the player), as well as added in the ability for the scripts to determine who did what and how to handle it

I rearranged some stuff for future work.

I had to zip one of the files and separate it into pieces using 7zip. The files are the Morak fbx and meta files located in Assets\Characters\Enemies\Morak.

To extract just use an unzipping tool on the first file... Ridiculous that one mixamo fbx and avater is 22 mb while the other is 117 mb